Use new http patterns and accept wildcards for existing methods (#2)
Some methods were split into getXX and postXX because of it. Actually, I could not to touch them, but it's more convenient to have the same approach for every handler.
Test every pattern carefully, please. I tested some, but not all of them. `test-web.sh` doesn't work for some reason.

Reviewed-on: https://codeberg.org/bouncepaw/betula/pulls/2
Co-authored-by: Danila Gorelko <danila@danilax86.space>
Co-committed-by: Danila Gorelko <danila@danilax86.space>