Use betula user agent when sending outgoing requests
Hi Timur,

glad to see the contribution is appreciated. I have added an additional
patch to be applied on top of the first one that removes the code to
retrieve the version logic, since I don't have any strong opinion here.

As to why it was done the way it was:

> 1) The file storing the version is called .version_string.txt. I do
>    not see why it should be a hidden file, or a file to begin with.

It is a file so it can be referenced via `go:embed`. It does not need to
be a hidden file, but I thought this may be useful in not cluttering up
file listings of the directory.

> 2) What if it's empty or does not exist? Not everyone will call
>    go generate.

`go build` fails if `go:embed` is called with a non-existant file. It
will not fail if it is called with an outdated file, or a file
containing arbitrary content.

As for the rest, as I said, I don't have any strong opinions and I think
a user agent without version information is useful enough.

Hope this helps!