Add PWA registration with share target (#56)
This commit makes it possible to install Betula as a PWA and to share
pages with it. This makes it easier to create bookmarks on mobile
devices, where the bookmarklet is unusable (or at least hard to use).

To test this, I installed the PWA on my android phone and shared some pages
with it. It seems to work fine on my end, but I'm not sure if it works on iOS devices.

Reviewed-on: https://codeberg.org/bouncepaw/betula/pulls/56
Co-authored-by: Guilherme Puida Moreira <guilherme@puida.xyz>
Co-committed-by: Guilherme Puida Moreira <guilherme@puida.xyz>