Notifications (#111) (#113)
Closes #111

A brutalist user notification system. Different from page-local error notifications and those misconfiguration notifications.

I have a different design in mind, but it is blocked by lack of the new reposts and userpics. For now, this is good enough, I like how it looks.

In this PR I'm trying the new architectural style with services and repos. The Betula architecture cannot change overnight, but at least new components can be in the new style. One day I hope everything will be written similarly. For demonstration, refactored most of archive-related work to that style.

Reviewed-on: https://codeberg.org/bouncepaw/betula/pulls/113
Co-authored-by: Timur Ismagilov <bouncepaw2@ya.ru>
Co-committed-by: Timur Ismagilov <bouncepaw2@ya.ru>